+++ Simple Building Simulator/weather.cpp
@@ -10,8 +10,8 @@ Weather::~Weather()
 {
 }
 
-Eigen::MatrixXf Weather::GetWeatherForecast(uint32 duration, uint16 time_step, uint8 num_zones, uint8 num_rooms) {
-	int n = (duration / time_step) + 1;
+Eigen::MatrixXf Weather::GetWeatherForecast(long int duration, int time_step, int num_zones, int num_rooms) {
+	long int n = (duration / time_step) + 1;
 	int total_rooms = num_zones * num_rooms;
 
 	Eigen::MatrixXf ext_temperature = Eigen::MatrixXf::Ones(n, total_rooms) * 5;
